daemon.c: make _fork_parent static inline again
The commit cf649082c7d4 ("remove forced gcc optimization") removed -O3
optimization specified in the code for the function _fork_parent, but at
the same time it removed the 'static inline' part of the function
definition. That change seems unintended and _fork_parent is not a part
of the libc interface. Make it static inline again.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>